Add/update stubs for pilot.
Playing with a tentative module structure, based on some other examples.
Not sure what the final layout will be, but I think it would be more
productive to just go ahead and build some working code, and then
refactor that into a sensible module structure after the bare-bones API
is running.